refactor: applyCoupon 함수 useCart 훅 내부로 이동
